+++ packages/components/organisms/f-content-cards/test-utils/component-objects/f-content-cards.selectors.js
@@ -1,2 +1,11 @@
-exports.CONTENT_CARDS_COMPONENT = '[data-test-id="home-promotion-1"]';
+exports.PROMOTION_CARD_1_COMPONENT = '[data-test-id="home-promotion-1"]';
 exports.INNER_CONTAINER = '[data-test-id="home-promotion-1--container"]';
+exports.HEADER = '[data-test-id="home-promotion-1--header"]';
+exports.LOGO = '[data-test-id="home-promotion-1--logo"]';
+exports.SUBTITLE = '[data-test-id="home-promotion-1--subtitle"]';
+
+exports.PROMOTION_CARD_2_COMPONENT = '[data-test-id="home-promotion-2"]';
+exports.CTA = '[data-test-id="home-promotion-2--cta"]';
+exports.TEXT = '[data-test-id="home-promotion-2--text"]';
+exports.TITLE = '[data-test-id="home-promotion-2--title"]';
+exports.IMAGE = '[data-test-id="home-promotion-2--image"]';